+++ packages/components/CHANGELOG.md
@@ -14,7 +14,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - disabled and selected states
   - `brand` prop
   - `treeBackgroundColor` util function
+- `ProgressCircular` component
 - `UnorderedList`, `OrderedList`
+
 
 ### Changed